refactor(api): déplacer les contrôleurs de commentaires d'action vers actionComment/
refers to https://trello.com/c/evsOelCU/2717
