+++ packages/api/server/controllers/actionComment/createComment/actionComment.createComment.route.ts
@@ -1,6 +1,6 @@
 import { type ApplicationWithCustomRoutes } from '#server/loaders/customRouteMethodsLoader';
-import validator from './action.createComment.validator';
-import controller from './action.createComment';
+import validator from './actionComment.createComment.validator';
+import controller from './actionComment.createComment';
 
 export default (app: ApplicationWithCustomRoutes): void => {
     app.customRoutes.post('/actions/:id/comments', controller, validator, {
